#9- 2751 문제 풀이 1차 시도
