+++ Algorithm/Algorithm/BOJ/main.swift
@@ -1,21 +1,28 @@
 //
-//  2751.swift
-//  수 정렬하기 2
+//  1463.swift
+//  1로 만들기
 //
 //  Created by 김인영 on 2023/09/17.
 //
 
 import Foundation
 
-let n = Int(readLine()!)!
-var array = [Int]()
+var input = Int(readLine()!)!
+var count = 0
 
-for i in 0...n-1 {
-    array.append(Int(readLine()!)!)
+func caculate(_ input: Int) -> Int {
+    if input % 3 == 0 {
+        return input / 3
+    } else if input % 2 == 0 {
+        return input / 2
+    } else {
+        return input - 1
+    }
 }
 
-array.sort()
-
-for i in 0...n-1 {
-    print(array[i])
+while input != 1 {
+    input = caculate(input)
+    count += 1
 }
+
+print(count)
